Fix Scroll Snap
Fixed the bug where scroll snap would trigger when jumping to a section using the navBar. This would cause the jump to be shorter than intended. Jumping is implemented using scrollIntoView which unfortunately brings elements into view long enough for the scroll snap to trigger. The fix was to effectively create a mutex to guard the jumping call.